# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Innoscripta API",
        "version": "0.1"
    },
    "paths": {
        "/api/articles": {
            "get": {
                "tags": [
                    "Article"
                ],
                "summary": "Retrieve articles",
                "description": "Listing all the articles based on search criteria.",
                "operationId": "00f26012a65da5676fbac4c06bf86f35",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Category ID to filter by",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "source",
                        "in": "query",
                        "description": "Source ID to filter by",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 129
                        }
                    },
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Keyword to search for in articles",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "CeeDee"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Date to filter by",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-12-26"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArticleResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrors"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/preferred-articles": {
            "get": {
                "tags": [
                    "Article"
                ],
                "summary": "Preferred articles",
                "description": "Listing all the user preferred articles.",
                "operationId": "807328f7cad0b13980bf9acc916f00f2",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArticleResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrors"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Article": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "CeeDee Lamb is out for the year."
                    },
                    "description": {
                        "type": "string",
                        "example": "The Cowboys receiver finishes the season with 101 receptions for 1,194 and six touchdowns."
                    },
                    "url": {
                        "type": "string",
                        "format": "url",
                        "example": "https://www.nbcsports.com"
                    },
                    "published_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-12-26 21:31:22"
                    },
                    "source": {
                        "$ref": "#/components/schemas/Source"
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "author": {
                        "$ref": "#/components/schemas/Author"
                    }
                },
                "type": "object"
            },
            "ArticleResponse": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Article"
                        }
                    },
                    "links": {
                        "$ref": "#/components/schemas/PaginationLinks"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/PaginationMeta"
                    }
                },
                "type": "object"
            },
            "Author": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Charean Williams"
                    }
                },
                "type": "object"
            },
            "Category": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "general"
                    }
                },
                "type": "object"
            },
            "PaginationLinks": {
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "api/articles?page=1"
                    },
                    "last": {
                        "type": "string",
                        "example": "api/articles?page=1"
                    },
                    "prev": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "links": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "url": {
                                    "type": "string",
                                    "example": null,
                                    "nullable": true
                                },
                                "label": {
                                    "type": "string",
                                    "example": "Previous"
                                },
                                "active": {
                                    "type": "boolean",
                                    "example": false
                                }
                            },
                            "type": "object"
                        }
                    },
                    "path": {
                        "type": "string",
                        "example": ""
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "to": {
                        "type": "integer",
                        "example": 1
                    },
                    "total": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "Source": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "NBCSports.com"
                    }
                },
                "type": "object"
            },
            "ValidationErrors": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Validation error message"
                    },
                    "errors": {
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}